feat(org-unit-tree): use dhis2/ui-widgets OrganisationUnitTree
