--- src/components/OrgUnitTree/OrgUnitTree.module.css
@@ -1,3 +1,0 @@
-.container {
-    margin-bottom: 15px;
-}
