+++ src/components/index.js
@@ -25,7 +25,6 @@ export { JobSummary } from './JobSummary/JobSummary'
 export { MiniJobSummary } from './JobSummary/MiniJobSummary/MiniJobSummary'
 export { MoreOptions } from './MoreOptions/MoreOptions'
 export { OrgUnitIdScheme } from './ElementSchemes/OrgUnitIdScheme'
-export { OrgUnitTree } from './OrgUnitTree/OrgUnitTree'
 export { OrgUnitTreeField } from './OrgUnitTree/OrgUnitTreeField'
 export { Page } from './Page/Page'
 export { ProgramStages } from './ProgramStages/ProgramStages'
